fix: move to different app structure
