+++ apps/build-onchain-apps/pages/buy-me-coffee/index.tsx
@@ -1,10 +1,10 @@
 import { useCallback } from 'react';
 import { Theme, Box, Container, Flex, Grid, Section, Text } from '@radix-ui/themes';
-import Header from '@/components/Header';
-import { TitleAndMetaTags } from '@/components/TitleAndMetaTags';
-import useOnchainCoffeeMemos from './hooks/useOnchainCoffeeMemos';
-import FormBuyCoffee from './components/FormBuyCoffee';
-import Memos from './components/Memos';
+import Header from '../../src/components/Header';
+import { TitleAndMetaTags } from '../../src/components/TitleAndMetaTags';
+import useOnchainCoffeeMemos from '../../src/hooks/useOnchainCoffeeMemos';
+import FormBuyCoffee from '../../src/components/buy-me-coffee/FormBuyCoffee';
+import Memos from '../../src/components/buy-me-coffee/Memos';
 
 export default function BuyMeCoffee() {
   const { memos, refetchMemos } = useOnchainCoffeeMemos();
