+++ apps/build-onchain-apps/pages/mint/index.tsx
@@ -1,12 +1,12 @@
 import { Theme, Container, Section } from '@radix-ui/themes';
 import dynamic from 'next/dynamic';
-import Header from '@/components/Header';
-import { TitleAndMetaTags } from '@/components/TitleAndMetaTags';
+import Header from '../../src/components/Header';
+import { TitleAndMetaTags } from '../../src/components/TitleAndMetaTags';
 
 // Because the mint page relies so heavily on client-side state, without disabling SSR
 // for its internals we get annoying hydration errors. A future enhancement would be to
 // read token metadata through a provider that is available server-side.
-const Mint = dynamic(async () => import('./components/Mint').then((mod) => mod), {
+const Mint = dynamic(async () => import('../../src/components/mint/Mint').then((mod) => mod), {
   ssr: false,
 });
