+++ apps/build-onchain-apps/src/components/buy-me-coffee/Memos.tsx
@@ -1,6 +1,6 @@
 import { Flex } from '@radix-ui/themes';
 import MemoCard from './MemoCard';
-import type { CoffeeMemo } from '@/types';
+import type { CoffeeMemo } from '../../types';
 
 type MemosProps = {
   memos: CoffeeMemo[];
