+++ apps/build-onchain-apps/src/hooks/useCollectionMetadata.ts
@@ -1,7 +1,7 @@
 import { useCallback, useEffect, useState } from 'react';
 import { useContractRead } from 'wagmi';
-import { CONTRACT_CUSTOM_1155 } from '@/onchain/contract/contractInfo';
-import ipfsToHTTP from '@/utils/ipfsToHTTP';
+import { CONTRACT_CUSTOM_1155 } from '../../src/onchain/contract/contractInfo';
+import ipfsToHTTP from '../../src/utils/ipfsToHTTP';
 
 // A future enhancement would be to track error state from the contract read
 // and the fetch so that we can gracefully surface issues to users.
