+++ apps/build-onchain-apps/src/styles/global.css
